requirements.txt
google-generativeai>=0.3.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0 
dataset